+++ src/speclj/running.clj
@@ -26,18 +26,27 @@
 
 (defn- eval-characteristic [befores body afters]
   (eval-components befores)
-  (body)
-  (eval-components afters))
+  (try
+    (body)
+    (finally
+      (eval-components afters))))
 
 (defn- reset-withs [withs]
   (doseq [with withs] (reset-with with)))
 
+(defn- collect-components [getter description]
+  (loop [description description components []]
+    (if description
+      (recur @(.parent description) (concat (getter description) components))
+      components)))
+
 (defn- do-characteristic [characteristic reporter]
   (let [description @(.description characteristic)
-        befores @(.befores description)
-        body (nested-fns (.body characteristic) (map #(.body %) @(.arounds description)))
-        afters @(.afters description)
-        withs @(.withs description)
+        befores (collect-components #(deref (.befores %)) description)
+        arounds (collect-components #(deref (.arounds %)) description)
+        body (nested-fns (.body characteristic) (map #(.body %) arounds))
+        afters (collect-components #(deref (.afters %)) description)
+        withs (collect-components #(deref (.withs %)) description)
         start-time (System/nanoTime)]
     (try
       (eval-characteristic befores body afters)
@@ -51,25 +60,33 @@
       (finally
         (reset-withs withs))))) ;MDM - Possible clojure bug.  Inlining reset-withs results in compile error 
 
-(defn- do-characteristics [characteristics description reporter]
+(defn- do-characteristics [characteristics reporter]
   (doall
     (for [characteristic characteristics]
       (do-characteristic characteristic reporter))))
 
+(defn- withs-mapping [description]
+  (let [withs @(.withs description)
+        ns (.ns description)]
+    (reduce #(assoc %1 (ns-resolve ns (.name %2)) %2) {} withs)))
+
 (defn do-description [description reporter]
   (report-description reporter description)
   (eval-components @(.before-alls description))
-  (let [results (do-characteristics @(.charcteristics description) description reporter)]
-    (eval-components @(.after-alls description))
-    results))
+  (with-bindings (withs-mapping description)
+    (let [results (do-characteristics @(.charcteristics description) reporter)]
+      (loop [results results descriptions @(.children description)]
+        (if (seq descriptions)
+          (recur (concat results (do-description (first descriptions) reporter)) (rest descriptions))
+          (do
+            (eval-components @(.after-alls description))
+            results))))))
 
 (defprotocol Runner
   (run-directories [this directories reporter])
+  (submit-description [this description])
   (run-description [this description reporter])
-  (report [this reporter]))
-
-(defn submit-description [description]
-  (run-description (active-runner) description (active-reporter)))
+  (run-and-report [this reporter]))
 
 (def clj-file-regex #".*\.clj")
 (defn clj-files-in [& dirs]