+++ CHANGES.md
@@ -1,7 +1,11 @@
-# TBD
+# 1.1.0
 
+* standard runner won't evaluate specs until they're all loaded
+* `with` components are only bound within their context
+* nested describe/context
+* `after`s are invoked even after failures or errors
 * should= support for doubles (use delta)
-* Lazy seqs print nicely in output (trptcolin)
+* Lazy seqs print nicely in output (@trptcolin)
 
 # 1.0.3